chore(preprocess-twmerge): break up files into smaller chunks docs(preprocess-twmerge): extensively document AST parsing
