+++ core/vite/plugins/preprocess-twmerge/src/ast/traverse-ast.ts
@@ -5,6 +5,7 @@ export function traverseAST(
   parent: Node | null = null,
 ) {
   if (!node || typeof node !== "object") return;
+  // FIXME: Create a symbol of the node so we can track its changes. This would allow us to avoid unnecessary re-evaluation of the expression, and to keep AST modifications in sync with the original source code.
   onVisit(node, parent);
 
   for (const value of Object.values(node)) {
