+++ core/vite/plugins/preprocess-twmerge/tsconfig.json
@@ -1,8 +1,8 @@
 {
   "extends": "../../../../tsconfig.base.json",
   "compilerOptions": {
-    "noEmit": false,
-    "allowImportingTsExtensions": false,
+    "emitDeclarationOnly": true,
+    "allowImportingTsExtensions": true,
     "paths": {
       "@/*": ["./src/*"]
     }
